+++ CHANGELOG.markdown
@@ -1,3 +1,7 @@
+next
+----
+* Added `Distributive` instances for `GHC.Generics` datatypes
+
 0.5.0.2
 -------
 * A more elegant fix for builds on GHC 7.2